Bump version to 0.2.0
