+++ stringen.py
@@ -1,4 +1,6 @@
-from stringen.__main__ import main
+"""Support running ``python stringen.py`` directly."""
+
+from stringen.cli import main
 
 if __name__ == "__main__":
     main()
